feat!: white/blacklist store state by keys, not filtering by mutation name
The previous process filtered mutations, but I don't think that was enough.
An initial mutation is filtered, but subsequent mutations could be allowed.
The main issue was that the entire state is committed to the electron-store.
Now, individual keys are committed.

BREAKING CHANGE: options for persisted-state have changed. whitelist and blacklist are now store property paths.